feat: Implement issue-1 (resart app / retry API) (#2)
This PR fixes #1

Running the app as a systemd service
Linux systemd services can be restarted automatically in the event of a crash.
Modify the README instructions to set up the app as a systemd service that starts at boot.

Retrying DREO APIs
Configure axios with a 4 seconds timeout to avoid connection issues with the server. Also integrate with axios-retry to automatically retry idempotent API calls that fail due to network issues or due to "connection aborted".